requirements.txt
aiogram==3.13.1
aiosqlite==0.19.0
python-dotenv==1.0.1
aiohttp==3.9.1
aiofiles==23.2.1

